Fixed data download and display of summary graph issue
Download option was not working as database moved to new server with 
limited privileges. Modified the php to execute from command line to 
generate and download csv file.
Add a common style for error message.
Fixed the AQI reading display and display error message for which no 
graph is plotted on summary page.